switched to relative imports, moved functional parts to lib for better usage as subproject
