+++ lib/data_generator/object_detection_2d_image_boxes_validation_utils.py
@@ -23,7 +23,7 @@ limitations under the License.
 from __future__ import division
 import numpy as np
 
-from bounding_box_utils.bounding_box_utils import iou
+from ..bounding_box_utils.bounding_box_utils import iou
 
 class BoundGenerator:
     '''
