+++ lib/data_generator/object_detection_2d_misc_utils.py
