+++ lib/data_generator/object_detection_2d_patch_sampling_ops.py
@@ -19,7 +19,7 @@ limitations under the License.
 from __future__ import division
 import numpy as np
 
-from data_generator.object_detection_2d_image_boxes_validation_utils import BoundGenerator, BoxFilter, ImageValidator
+from .object_detection_2d_image_boxes_validation_utils import BoundGenerator, BoxFilter, ImageValidator
 
 class PatchCoordinateGenerator:
     '''
