+++ lib/data_generator/object_detection_2d_photometric_ops.py
