+++ lib/keras_layers/keras_layer_AnchorBoxes.py
@@ -22,7 +22,7 @@ import tensorflow.keras.backend as K
 from tensorflow.keras.layers import InputSpec
 from tensorflow.keras.layers import Layer
 
-from bounding_box_utils.bounding_box_utils import convert_coordinates
+from ..bounding_box_utils.bounding_box_utils import convert_coordinates
 
 class AnchorBoxes(Layer):
     '''
