+++ lib/keras_layers/keras_layer_L2Normalization.py
