+++ lib/ssd_encoder_decoder/ssd_input_encoder.py
@@ -19,8 +19,8 @@ limitations under the License.
 from __future__ import division
 import numpy as np
 
-from bounding_box_utils.bounding_box_utils import iou, convert_coordinates
-from ssd_encoder_decoder.matching_utils import match_bipartite_greedy, match_multi
+from ..bounding_box_utils.bounding_box_utils import iou, convert_coordinates
+from .matching_utils import match_bipartite_greedy, match_multi
 
 class SSDInputEncoder:
     '''
