+++ lib/ssd_encoder_decoder/ssd_output_decoder.py
@@ -22,7 +22,7 @@ limitations under the License.
 from __future__ import division
 import numpy as np
 
-from bounding_box_utils.bounding_box_utils import iou, convert_coordinates
+from ..bounding_box_utils.bounding_box_utils import iou, convert_coordinates
 
 def greedy_nms(y_pred_decoded, iou_threshold=0.45, coords='corners', border_pixels='half'):
     '''
